healthupdate location update and webgl builds
