+++ Assets/Scenes/Level_0.unity
@@ -153,7 +153,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5384936018756839015, guid: d7d18dd0e2b024e99888862594d3e711, type: 3}
       propertyPath: Duration
-      value: 60
+      value: 120
       objectReference: {fileID: 0}
     - target: {fileID: 5384936018756839015, guid: d7d18dd0e2b024e99888862594d3e711, type: 3}
       propertyPath: gameOverHUD
