+++ Assets/enemies/enemyBulletYellow.prefab
@@ -188,4 +188,4 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 73246de04323f4a3b8574cbeb52b5934, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  projectileSpeed: 15
+  projectileSpeed: 10
