--- Assets/healthUpdate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 09f7666ab5adc43dab9ae0c80f0d3e35
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
