+++ ProjectSettings/Packages/com.unity.connect.share/Settings.json
@@ -6,7 +6,7 @@
             {
                 "type": "System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089",
                 "key": "buildOutputDirList",
-                "value": "{\"m_Value\":\"C:/UNITY Projects/CSCI526/csci526-professional-nappers/WegGLBuilds;;;;;;;;;\"}"
+                "value": "{\"m_Value\":\"C:/UNITY Projects/CSCI526/csci526-professional-nappers/WegGLBuilds/GreyBoxPrototype27Sept2022;C:/UNITY Projects/CSCI526/csci526-professional-nappers/WegGLBuilds;;;;;;;;\"}"
             },
             {
                 "type": "System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089",
