+++ WegGLBuilds/GreyBoxPrototype27Sept2022/ProjectVersion.txt
@@ -1,0 +1,2 @@
+m_EditorVersion: 2021.3.9f1
+m_EditorVersionWithRevision: 2021.3.9f1 (ad3870b89536)
