+++ WegGLBuilds/GreyBoxPrototype27Sept2022/StreamingAssets/UnityServicesProjectConfiguration.json
@@ -1,0 +1,1 @@
+{"Keys":["com.unity.services.core.cloud-environment","com.unity.services.core.version"],"Values":[{"m_Value":"production","m_IsReadOnly":true},{"m_Value":"1.4.2","m_IsReadOnly":true}]}
