+++ WegGLBuilds/GreyBoxPrototype27Sept2022Draft2/dependencies.txt
@@ -1,0 +1,46 @@
+com.unity.cinemachine@2.8.9
+com.unity.collab-proxy@1.17.2
+com.unity.connect.share@4.2.3
+com.unity.feature.development@1.0.1
+com.unity.ide.rider@3.0.15
+com.unity.ide.visualstudio@2.0.16
+com.unity.ide.vscode@1.2.5
+com.unity.inputsystem@1.4.2
+com.unity.probuilder@5.0.6
+com.unity.test-framework@1.1.31
+com.unity.textmeshpro@3.0.6
+com.unity.timeline@1.6.4
+com.unity.ugui@1.0.0
+com.unity.visualscripting@1.7.8
+com.unity.modules.ai@1.0.0
+com.unity.modules.androidjni@1.0.0
+com.unity.modules.animation@1.0.0
+com.unity.modules.assetbundle@1.0.0
+com.unity.modules.audio@1.0.0
+com.unity.modules.cloth@1.0.0
+com.unity.modules.director@1.0.0
+com.unity.modules.imageconversion@1.0.0
+com.unity.modules.imgui@1.0.0
+com.unity.modules.jsonserialize@1.0.0
+com.unity.modules.particlesystem@1.0.0
+com.unity.modules.physics@1.0.0
+com.unity.modules.physics2d@1.0.0
+com.unity.modules.screencapture@1.0.0
+com.unity.modules.terrain@1.0.0
+com.unity.modules.terrainphysics@1.0.0
+com.unity.modules.tilemap@1.0.0
+com.unity.modules.ui@1.0.0
+com.unity.modules.uielements@1.0.0
+com.unity.modules.umbra@1.0.0
+com.unity.modules.unityanalytics@1.0.0
+com.unity.modules.unitywebrequest@1.0.0
+com.unity.modules.unitywebrequestassetbundle@1.0.0
+com.unity.modules.unitywebrequestaudio@1.0.0
+com.unity.modules.unitywebrequesttexture@1.0.0
+com.unity.modules.unitywebrequestwww@1.0.0
+com.unity.modules.vehicles@1.0.0
+com.unity.modules.video@1.0.0
+com.unity.modules.vr@1.0.0
+com.unity.modules.wind@1.0.0
+com.unity.modules.xr@1.0.0
+com.unity.template.3d@8.1.0
